final/requirements.txt
streamlit
pandas
google-genai
requests